fix RSBaseNodeCommand permission
🤖 AI[0%] 🔧 Human Fixed[0%] 🧑 Human[100%] 👌 AI Adopted[0%]

+++ rosen/modules/render_service/main/render_server/transaction/zidl/rs_client_to_service_connection_stub.cpp
@@ -1197,7 +1197,7 @@ int RSClientToServiceConnectionStub::OnRemoteRequest(
                 break;
             }
             int32_t result = GetDisplayIdentificationData(id, outPort, edidData);
-            if (!reply.WriteUint8(result)) {
+            if (!reply.WriteInt32(result)) {
                 RS_LOGE("RSClientToServiceConnectionStub::GET_DISPLAY_IDENTIFICATION_DATA Write result failed!");
                 ret = IPC_STUB_WRITE_PARCEL_ERR;
                 break;
